Add files via upload
I volunteer annually for New England Give Camp as a technical developer https://newenglandgivecamp.org/, https://www.facebook.com/NEGiveCamp/, https://www.facebook.com/NEGiveCamp/videos/10155164927309603/.  At the conclusion of the weekend hackathon, we give professional presenatations on our work.  I put this together for the 2015 year when I worked on the team for Mystic River Watershed.  Unfortunately, I was not able to get screenshots of the new website in our presentation, but it can be seen at this URL: https://mysticriver.org/